refactor: extract some function from ImgurPlugin class
